Made a generic `AnnotatorPlugin` component
